threadbackend/requirements.txt
appnope==0.1.0
beautifulsoup4==4.6.0
bleach==2.1.2
bs4==0.0.1
cachetools==3.1.0
certifi==2018.11.29
chardet==3.0.4
cycler==0.10.0
decorator==4.2.1
Django==2.1.7
django-tinymce==2.8.0
Django-tinymce-filebrowser==0.2.1
django-tinymce4-lite==1.7.5
docutils==0.14
entrypoints==0.2.3
html5lib==1.0.1
idna==2.8
ipykernel==4.8.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.0
jedi==0.11.1
Jinja2==2.10
jsmin==2.2.2
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.1
jupyter-console==5.2.0
jupyter-core==4.4.0
Kivy==1.10.1.dev0
Kivy-Garden==0.1.4
MarkupSafe==1.0
matplotlib==2.1.2
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
nltk==3.2.5
notebook==5.3.1
numpy==1.14.0
pandocfilters==1.4.2
parso==0.1.1
pexpect==4.3.1
pickleshare==0.7.4
Pillow==5.4.1
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
python-devtools==2
python-Levenshtein==0.11.2
pytz==2017.3
PyX==0.14.1
pyzmq==16.0.3
qtconsole==4.3.1
requests==2.21.0
scikit-learn==0.18
scipy==0.16.0
Send2Trash==1.4.2
simplegeneric==0.8.1
six==1.11.0
sorl-thumbnail==12.5.0
terminado==0.8.1
testpath==0.3.1
tornado==4.5.3
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.1.0
